Factor common action test into a lib
